Eveapi module split into more files, added tests, readme
