--- __init__.py
@@ -1,3 +1,0 @@
-from eveapi import (AuthenticationError, Error, RequestError, ServerError,
-                    EVEAPIConnection)
-
