+++ tests/api.py
@@ -3,7 +3,7 @@
 #=============================================================================
 #
 # This file is in the Public Domain - Do with it as you please.
-# 
+#
 # THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
 # EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
 # OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -15,16 +15,21 @@
 #
 #----------------------------------------------------------------------------
 # Put your userID and apiKey (full access) here before running this script.
-YOUR_KEYID = 123456
-YOUR_VCODE = "nyanyanyanyanyanyanyanyanyanyanyanyanyanyanyanyanyanya:3"
+YOUR_KEYID = 2084022
+YOUR_VCODE = "4mdjugfRfv0SbdNT9cbpistAdm6UZjaOp09bllpbYCmgu3oB7LjfoxVllKoYA5ku"
 
-import time
+import datetime
 import tempfile
 import cPickle
 import zlib
+import sys
 import os
 from os.path import join, exists
-from httplib import HTTPException
+
+# This is just for correct path setting to eveapi folder, only required during
+# testing.
+eveapi_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), '..')
+sys.path.insert(0, eveapi_path)
 
 import eveapi
 
@@ -50,9 +55,9 @@ result1 = api.eve.AllianceList()
 # we'll just iterate over it and display all alliances with more than 1000
 # members:
 for alliance in result1.alliances:
-	if alliance.memberCount >= 1000:
-		print "%s <%s> has %d members" %\
-			(alliance.name, alliance.shortName, alliance.memberCount)
+    if alliance.memberCount >= 1000:
+        print "%s <%s> has %d members" %\
+            (alliance.name, alliance.shortName, alliance.memberCount)
 
 
 #-----------------------------------------------------------------------------
@@ -85,13 +90,13 @@ rich_charID = 0
 # Now the best way to iterate over the characters on your account and show
 # the isk balance is probably this way:
 for character in result2.characters:
-	wallet = auth.char.AccountBalance(characterID=character.characterID)
-	isk = wallet.accounts[0].balance
-	print character.name, "has", isk, "ISK."
+    wallet = auth.char.AccountBalance(characterID=character.characterID)
+    isk = wallet.accounts[0].balance
+    print character.name, "has", isk, "ISK."
 
-	if isk > rich:
-		rich = isk
-		rich_charID = character.characterID
+    if isk > rich:
+        rich = isk
+        rich_charID = character.characterID
 
 
 
@@ -118,15 +123,15 @@ print
 #
 
 try:
-	# Try calling account/Characters without authentication context
-	api.account.Characters()
+    # Try calling account/Characters without authentication context
+    api.account.Characters()
 except eveapi.Error, e:
-	print "Oops! eveapi returned the following error:"
-	print "code:", e.code
-	print "message:", e.message
+    print "Oops! eveapi returned the following error:"
+    print "code:", e.code
+    print "message:", e.message
 except Exception, e:
-	print "Something went horribly wrong:", str(e)
-	raise
+    print "Something went horribly wrong:", str(e)
+    raise
 
 
 #-----------------------------------------------------------------------------
@@ -172,26 +177,26 @@ total_skills = 0
 # tree...
 for g in skilltree.skillGroups:
 
-	skills_trained_in_this_group = False
+    skills_trained_in_this_group = False
 
-	# ... iterate over the skills in this group...
-	for skill in g.skills:
+    # ... iterate over the skills in this group...
+    for skill in g.skills:
 
-		# see if we trained this skill by checking the character sheet object
-		trained = sheet.skills.Get(skill.typeID, False)
-		if trained:
-			# yep, we trained this skill.
+        # see if we trained this skill by checking the character sheet object
+        trained = sheet.skills.Get(skill.typeID, False)
+        if trained:
+            # yep, we trained this skill.
 
-			# print the group name if we haven't done so already
-			if not skills_trained_in_this_group:
-				print g.groupName
-				skills_trained_in_this_group = True
+            # print the group name if we haven't done so already
+            if not skills_trained_in_this_group:
+                print g.groupName
+                skills_trained_in_this_group = True
 
-			# and display some info about the skill!
-			print "- %s Rank(%d) - SP: %d/%d - Level: %d" %\
-				(skill.typeName, skill.rank, trained.skillpoints, (skill.rank * sp[trained.level]), trained.level)
-			total_skills += 1
-			total_sp += trained.skillpoints
+            # and display some info about the skill!
+            print "- %s Rank(%d) - SP: %d/%d - Level: %d" %\
+                (skill.typeName, skill.rank, trained.skillpoints, (skill.rank * sp[trained.level]), trained.level)
+            total_skills += 1
+            total_sp += trained.skillpoints
 
 
 # And to top it off, display totals.
@@ -217,11 +222,11 @@ rowset.SortBy("shortName")
 # Note the use of Select() here. The Select method speeds up iterating over
 # large rowsets considerably as no temporary row instances are created.
 for ticker in rowset.Select("shortName"):
-	print ticker,
+    print ticker,
 print
 
 # The sort above modified the result inplace. There is another method, called
-# SortedBy, which returns a new rowset. 
+# SortedBy, which returns a new rowset.
 
 print
 
@@ -240,9 +245,9 @@ alliances_by_ticker = rowset.IndexedBy("shortName")
 # Assumes ISD alliance exists. If it doesn't, we probably have bigger
 # problems than the unhandled exception here -_-
 try:
-	print alliances_by_ticker.Get("ISD")
+    print alliances_by_ticker.Get("ISD")
 except :
-	print "Blimey! CCP let the ISD alliance expire -AGAIN-. How inconvenient!"
+    print "Blimey! CCP let the ISD alliance expire -AGAIN-. How inconvenient!"
 
 # You may specify a default to return in case the row wasn't found:
 print alliances_by_ticker.Get("123456", 42)
@@ -250,9 +255,9 @@ print alliances_by_ticker.Get("123456", 42)
 # If no default was specified and you try to look up a key that does not
 # exist, an appropriate exception will be raised:
 try:
-	print alliances_by_ticker.Get("123456")
+    print alliances_by_ticker.Get("123456")
 except KeyError:
-	print "This concludes example 5"
+    print "This concludes example 5"
 
 
 
@@ -263,80 +268,80 @@ print
 
 # For some calls you will want caching. To facilitate this, a customized
 # cache handler can be attached. Below is an example of a simple cache
-# handler. 
+# handler.
 
 class MyCacheHandler(object):
-	# Note: this is an example handler to demonstrate how to use them.
-	# a -real- handler should probably be thread-safe and handle errors
-	# properly (and perhaps use a better hashing scheme).
+    # Note: this is an example handler to demonstrate how to use them.
+    # a -real- handler should probably be thread-safe and handle errors
+    # properly (and perhaps use a better hashing scheme).
 
-	def __init__(self, debug=False):
-		self.debug = debug
-		self.count = 0
-		self.cache = {}
-		self.tempdir = join(tempfile.gettempdir(), "eveapi")
-		if not exists(self.tempdir):
-			os.makedirs(self.tempdir)
+    def __init__(self, debug=False):
+        self.debug = debug
+        self.count = 0
+        self.cache = {}
+        self.tempdir = join(tempfile.gettempdir(), "eveapi")
+        if not exists(self.tempdir):
+            os.makedirs(self.tempdir)
 
-	def log(self, what):
-		if self.debug:
-			print "[%d] %s" % (self.count, what)
+    def log(self, what):
+        if self.debug:
+            print "[%d] %s" % (self.count, what)
 
-	def retrieve(self, host, path, params):
-		# eveapi asks if we have this request cached
-		key = hash((host, path, frozenset(params.items())))
+    def retrieve(self, host, path, params):
+        # eveapi asks if we have this request cached
+        key = hash((host, path, frozenset(params.items())))
 
-		self.count += 1  # for logging
+        self.count += 1  # for logging
 
-		# see if we have the requested page cached...
-		cached = self.cache.get(key, None)
-		if cached:
-			cacheFile = None
-			#print "'%s': retrieving from memory" % path
-		else:
-			# it wasn't cached in memory, but it might be on disk.
-			cacheFile = join(self.tempdir, str(key) + ".cache")
-			if exists(cacheFile):
-				self.log("%s: retrieving from disk" % path)
-				f = open(cacheFile, "rb")
-				cached = self.cache[key] = cPickle.loads(zlib.decompress(f.read()))
-				f.close()
+        # see if we have the requested page cached...
+        cached = self.cache.get(key, None)
+        if cached:
+            cacheFile = None
+            #print "'%s': retrieving from memory" % path
+        else:
+            # it wasn't cached in memory, but it might be on disk.
+            cacheFile = join(self.tempdir, str(key) + ".cache")
+            if exists(cacheFile):
+                self.log("%s: retrieving from disk" % path)
+                f = open(cacheFile, "rb")
+                cached = self.cache[key] = cPickle.loads(zlib.decompress(f.read()))
+                f.close()
 
-		if cached:
-			# check if the cached doc is fresh enough
-			if time.time() < cached[0]:
-				self.log("%s: returning cached document" % path)
-				return cached[1]  # return the cached XML doc
+        if cached:
+            # check if the cached doc is fresh enough
+            if datetime.datetime.now() < cached[0]:
+                self.log("%s: returning cached document" % path)
+                return cached[1]  # return the cached XML doc
 
-			# it's stale. purge it.
-			self.log("%s: cache expired, purging!" % path)
-			del self.cache[key]
-			if cacheFile:
-				os.remove(cacheFile)
+            # it's stale. purge it.
+            self.log("%s: cache expired, purging!" % path)
+            del self.cache[key]
+            if cacheFile:
+                os.remove(cacheFile)
 
-		self.log("%s: not cached, fetching from server..." % path)
-		# we didn't get a cache hit so return None to indicate that the data
-		# should be requested from the server.
-		return None
+        self.log("%s: not cached, fetching from server..." % path)
+        # we didn't get a cache hit so return None to indicate that the data
+        # should be requested from the server.
+        return None
 
-	def store(self, host, path, params, doc, obj):
-		# eveapi is asking us to cache an item
-		key = hash((host, path, frozenset(params.items())))
+    def store(self, host, path, params, doc, obj):
+        # eveapi is asking us to cache an item
+        key = hash((host, path, frozenset(params.items())))
 
-		cachedFor = obj.cachedUntil - obj.currentTime
-		if cachedFor:
-			self.log("%s: cached (%d seconds)" % (path, cachedFor))
+        cachedFor = obj.cachedUntil - obj.currentTime
+        if cachedFor:
+            self.log("%s: cached (%s seconds)" % (path, cachedFor.seconds))
 
-			cachedUntil = time.time() + cachedFor
+            cachedUntil = datetime.datetime.now() + cachedFor
 
-			# store in memory
-			cached = self.cache[key] = (cachedUntil, doc)
+            # store in memory
+            cached = self.cache[key] = (cachedUntil, doc)
 
-			# store in cache folder
-			cacheFile = join(self.tempdir, str(key) + ".cache")
-			f = open(cacheFile, "wb")
-			f.write(zlib.compress(cPickle.dumps(cached, -1)))
-			f.close()
+            # store in cache folder
+            cacheFile = join(self.tempdir, str(key) + ".cache")
+            f = open(cacheFile, "wb")
+            f.write(zlib.compress(cPickle.dumps(cached, -1)))
+            f.close()
 
 
 # Now try out the handler! Even though were initializing a new api object
@@ -369,7 +374,7 @@ print
 # contexts in the chain anyway, this is okay.
 me = cachedApi.auth(keyID=YOUR_KEYID, vCode=YOUR_VCODE).character(rich_charID)
 
-# Now fetch the journal. Since this character context was created through 
+# Now fetch the journal. Since this character context was created through
 # the cachedApi object, it will still use the cachehandler from example 5.
 journal = me.WalletJournal()
 
@@ -391,26 +396,25 @@ entriesByRefType = journal.transactions.GroupedBy("refTypeID")
 amount = 0.0
 date = 0
 for taxAmount, date in entriesByRefType[54].Select("amount", "date"):
-	amount += -taxAmount
+    amount += -taxAmount
 
 print "You paid a %.2f ISK transaction tax since %s" %\
-	(amount, time.asctime(time.gmtime(date)))
+    (amount, date)
 
 
 # You might also want to see how much a certain item yielded you recently.
-typeName = "Expanded Cargohold II"  # change this to something you sold.
+typeName = "Hammerhead I"  # change this to something you sold.
 amount = 0.0
 
 wallet = me.WalletTransactions()
 soldTx = wallet.transactions.GroupedBy("transactionType")["sell"]
 for row in soldTx.GroupedBy("typeName")[typeName]:
-	amount += (row.quantity * row.price)
+    amount += (row.quantity * row.price)
 
 print "%s sales yielded %.2f ISK since %s" %\
-	(typeName, amount, time.asctime(time.gmtime(row.transactionDateTime)))
+    (typeName, amount, row.transactionDateTime)
 
 # I'll leave walking the transaction pages as an excercise to the reader ;)
 # Please also see the eveapi module itself for more documentation.
 
 # That's all folks!
-
